+++ @crosshatch/util/Env.ts
@@ -1,4 +1,4 @@
-import { Context, Config, Effect, Option, Schema as S } from "effect"
+import { Context, Config, Effect, Option, Schema as S, Layer } from "effect"
 
 export type Stage = typeof Stage.Type
 export const Stage = S.Union([
@@ -12,17 +12,23 @@ export interface UrlConfig {
   readonly pathname?: string | undefined
 }
 
-export class ChxEnv extends Context.Reference<{
-  readonly stage: Stage
-  readonly domain: (config?: UrlConfig) => string
-  readonly url: (config?: UrlConfig) => string
-}>("crosshatch/ChxEnv", {
-  defaultValue: () =>
+export class Env extends Context.Service<
+  Env,
+  {
+    readonly stage: Stage
+    readonly domain: (config?: UrlConfig) => string
+    readonly url: (config?: UrlConfig) => string
+  }
+>()("crosshatch/Env") {}
+
+export const layerFromHostname = (hostname: string) =>
+  Layer.effect(
+    Env,
     Effect.gen(function* () {
       const raw =
-        (import.meta as { readonly env?: undefined | { readonly VITE_PUBLIC_ALCHEMY_STAGE?: string | undefined } }).env
-          ?.VITE_PUBLIC_ALCHEMY_STAGE ??
-        (yield* Config.string("ALCHEMY_STAGE").pipe(
+        (import.meta as { readonly env?: undefined | { readonly VITE_PUBLIC_STAGE?: string | undefined } }).env
+          ?.VITE_PUBLIC_STAGE ??
+        (yield* Config.string("VITE_PUBLIC_STAGE").pipe(
           Config.option,
           Config.map(Option.getOrUndefined),
           Effect.catchTags({
@@ -32,9 +38,9 @@ export class ChxEnv extends Context.Reference<{
       const stage = yield* S.decodeUnknownEffect(Stage)(raw).pipe(Effect.orElseSucceed(() => "prod" as const))
       const domain = (config?: UrlConfig) => {
         const { sub, pathname } = config ?? {}
-        return `${stage.startsWith("staging-") ? `${stage}.` : ""}${sub ? `${sub}.` : ""}crosshatch.dev${stage.startsWith("dev_") ? ".localhost" : ""}${pathname ? `/${pathname}` : ""}`
+        return `${stage.startsWith("staging-") ? `${stage}.` : ""}${sub ? `${sub}.` : ""}${hostname}${stage.startsWith("dev_") ? ".localhost" : ""}${pathname ? `/${pathname}` : ""}`
       }
       const url = (config?: UrlConfig) => `https://${domain(config)}`
       return { stage, domain, url }
-    }).pipe(Effect.runSync),
-}) {}
+    }),
+  )
