+++ crosshatch/Cirque/Cirque.ts
@@ -1,2 +1,2 @@
 export * from "./CirqueApi.ts"
-export * from "./CirqueClient.ts"
+export * as CirqueClient from "./CirqueClient.ts"
