+++ crosshatch/Cli/main.ts
@@ -6,6 +6,7 @@ import { Command } from "effect/unstable/cli"
 
 import * as ChxNodeServices from "../ChxNodeServices/ChxNodeServices.ts"
 import { CirqueClient } from "../Cirque/Cirque.ts"
+import * as Env from "../Env.ts"
 import PackageJson from "../package.json" with { type: "json" }
 import { dev } from "./dev/dev.ts"
 import { facilitator } from "./facilitator/facilitator.ts"
@@ -19,7 +20,7 @@ Command.make("crosshatch").pipe(
   Effect.scoped,
   Effect.provide(
     Layer.mergeAll(
-      CirqueClient.layer.pipe(Layer.provideMerge(NodeHttpClient.layerFetch)),
+      CirqueClient.layer.pipe(Layer.provideMerge([NodeHttpClient.layerFetch, Env.layerFromHostname("crosshatch.dev")])),
       ChxNodeServices.layer.pipe(Layer.provideMerge(NodeServices.layer)),
     ),
   ),
