+++ crosshatch/index.ts
@@ -3,7 +3,7 @@ export * as AccountId from "./AccountId.ts"
 export * as Keychain from "./Keychain.ts"
 export * as Amount from "./Amount.ts"
 export * as Payer from "./Payer.ts"
-export * as ChxEnv from "./ChxEnv.ts"
+export * as Env from "@crosshatch/util/env"
 export * as Requirements from "./Requirements.ts"
 export * as ChainId from "./ChainId.ts"
 export * as Asset from "./Asset.ts"
